Revise scale; move Foursquare to it's own servie to be replace by Azure
